# SSE latin1 => utf32