--- a/src/westmere/sse_convert_latin1_to_utf32.cpp
+++ b/src/westmere/sse_convert_latin1_to_utf32.cpp
@@ -3,7 +3,7 @@ std::pair<const char*, char32_t*> sse_convert_latin1_to_utf32(const char* buf, s
 
     while (buf + 16 <= end) {
         // Load 16 Latin1 characters (16 bytes) into a 128-bit register
-        __m128i in = _mm_load_si128((__m128i*)buf);
+        __m128i in = _mm_loadu_si128((__m128i*)buf);
 
         // Shift input to process next 4 bytes
         __m128i in_shifted1 = _mm_srli_si128(in, 4);
@@ -16,10 +16,10 @@ std::pair<const char*, char32_t*> sse_convert_latin1_to_utf32(const char* buf, s
         __m128i out3 = _mm_cvtepu8_epi32(in_shifted2);
         __m128i out4 = _mm_cvtepu8_epi32(in_shifted3);
 
-        _mm_store_si128((__m128i*)utf32_output, out1);
-        _mm_store_si128((__m128i*)(utf32_output + 4), out2);
-        _mm_store_si128((__m128i*)(utf32_output + 8), out3);
-        _mm_store_si128((__m128i*)(utf32_output + 12), out4);
+        _mm_storeu_si128((__m128i*)utf32_output, out1);
+        _mm_storeu_si128((__m128i*)(utf32_output + 4), out2);
+        _mm_storeu_si128((__m128i*)(utf32_output + 8), out3);
+        _mm_storeu_si128((__m128i*)(utf32_output + 12), out4);
 
         utf32_output += 16;
         buf += 16;