For reference these are the benchmarks for the scalar:


convert_latin1_to_utf32+haswell, input size: 82168, iterations: 10000, dataset: unicode_lipsum/wikipedia_mars/esperanto.latin1.txt
  13.004 ins/byte,    7.311 cycle/byte,    0.423 GB/s (2.5 %),     3.094 GHz,    1.779 ins/cycle
  13.004 ins/char,    7.311 cycle/char,    0.423 Gc/s (2.5 %)     1.00 byte/char
input detected as unknown
current system detected as icelake

convert_latin1_to_utf32+haswell, input size: 432305, iterations: 10000, dataset: unicode_lipsum/wikipedia_mars/french.latin1.txt
  13.001 ins/byte,    7.464 cycle/byte,    0.427 GB/s (3.4 %),     3.184 GHz,    1.742 ins/cycle
  13.001 ins/char,    7.464 cycle/char,    0.427 Gc/s (3.4 %)     1.00 byte/char
input detected as unknown
current system detected as icelake

convert_latin1_to_utf32+haswell, input size: 199331, iterations: 10000, dataset: unicode_lipsum/wikipedia_mars/german.latin1.txt
  13.002 ins/byte,    7.420 cycle/byte,    0.417 GB/s (1.7 %),     3.093 GHz,    1.752 ins/cycle
  13.002 ins/char,    7.420 cycle/char,    0.417 Gc/s (1.7 %)     1.00 byte/char
input detected as unknown
current system detected as icelake

convert_latin1_to_utf32+haswell, input size: 271743, iterations: 10000, dataset: unicode_lipsum/wikipedia_mars/portuguese.latin1.txt
  13.001 ins/byte,    7.487 cycle/byte,    0.424 GB/s (2.6 %),     3.176 GHz,    1.736 ins/cycle
  13.001 ins/char,    7.487 cycle/char,    0.424 Gc/s (2.6 %)     1.00 byte/char
[leorio@localhost ~]$


This is the SSE version:


convert_latin1_to_utf32+westmere, input size: 82168, iterations: 10000, dataset: unicode_lipsum/wikipedia_mars/esperanto.latin1.txt
   4.944 ins/byte,    6.216 cycle/byte,    0.498 GB/s (0.4 %),     3.094 GHz,    0.795 ins/cycle
   4.944 ins/char,    6.216 cycle/char,    0.498 Gc/s (0.4 %)     1.00 byte/char
input detected as unknown
current system detected as icelake

convert_latin1_to_utf32+westmere, input size: 432305, iterations: 10000, dataset: unicode_lipsum/wikipedia_mars/french.latin1.txt
   4.939 ins/byte,    6.214 cycle/byte,    0.498 GB/s (0.3 %),     3.093 GHz,    0.795 ins/cycle
   4.939 ins/char,    6.214 cycle/char,    0.498 Gc/s (0.3 %)     1.00 byte/char
input detected as unknown
current system detected as icelake

convert_latin1_to_utf32+westmere, input size: 199331, iterations: 10000, dataset: unicode_lipsum/wikipedia_mars/german.latin1.txt
   4.940 ins/byte,    6.206 cycle/byte,    0.498 GB/s (0.3 %),     3.093 GHz,    0.796 ins/cycle
   4.940 ins/char,    6.206 cycle/char,    0.498 Gc/s (0.3 %)     1.00 byte/char
input detected as unknown
current system detected as icelake

convert_latin1_to_utf32+westmere, input size: 271743, iterations: 10000, dataset: unicode_lipsum/wikipedia_mars/portuguese.latin1.txt
   4.940 ins/byte,    6.209 cycle/byte,    0.498 GB/s (0.3 %),     3.093 GHz,    0.796 ins/cycle
   4.940 ins/char,    6.209 cycle/char,    0.498 Gc/s (0.3 %)     1.00 byte/char 

So not much faster, I'll see if I can squeeze a bit more performance out of it.