@lemire 2023-08-10
The drop in instructions is good. But the instructions per cycle is low.