@lemire 2023-08-10
@Nick-Nuon 

You process 16 characters with the following routine...

```
..B1.3:                         # Preds ..B1.1 ..B1.3
        movdqu    xmm2, XMMWORD PTR [rdi]                       #16.48
        movdqa    xmm0, xmm2                                    #22.31
        movdqa    xmm1, xmm2                                    #26.31
        pmovzxbd  xmm3, xmm2                                    #19.24
        mov       rdi, rax                                      #39.9
        psrldq    xmm0, 4                                       #22.31
        add       rax, 16                                       #14.12
        psrldq    xmm1, 8                                       #26.31
        psrldq    xmm2, 12                                      #30.31
        pmovzxbd  xmm0, xmm0                                    #23.24
        pmovzxbd  xmm1, xmm1                                    #27.24
        pmovzxbd  xmm2, xmm2                                    #31.24
        movdqu    XMMWORD PTR [rdx], xmm3                       #33.36
        movdqu    XMMWORD PTR [16+rdx], xmm0                    #34.37
        movdqu    XMMWORD PTR [32+rdx], xmm1                    #35.37
        movdqu    XMMWORD PTR [48+rdx], xmm2                    #36.37
        add       rdx, 64                                       #38.9
        cmp       rax, rsi                                      #14.24
        jbe       ..B1.3    
```

I don't know how you arrive at 4.944 ins/byte. I see fewer than 20 instructions processing 16 characters, so you should have slightly more than one instruction per character... ???

On something like Ice Lake,  [this should require about 4 cycles](https://uica.uops.info/?code=..B1.3%3A%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%23%20Preds%20..B1.1%20..B1.3%0D%0A%20%20%20%20%20%20%20%20movdqu%20%20%20%20xmm2%2C%20XMMWORD%20PTR%20%5Brdi%5D%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2316.48%0D%0A%20%20%20%20%20%20%20%20movdqa%20%20%20%20xmm0%2C%20xmm2%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2322.31%0D%0A%20%20%20%20%20%20%20%20movdqa%20%20%20%20xmm1%2C%20xmm2%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2326.31%0D%0A%20%20%20%20%20%20%20%20pmovzxbd%20%20xmm3%2C%20xmm2%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2319.24%0D%0A%20%20%20%20%20%20%20%20mov%20%20%20%20%20%20%20rdi%2C%20rax%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2339.9%0D%0A%20%20%20%20%20%20%20%20psrldq%20%20%20%20xmm0%2C%204%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2322.31%0D%0A%20%20%20%20%20%20%20%20add%20%20%20%20%20%20%20rax%2C%2016%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2314.12%0D%0A%20%20%20%20%20%20%20%20psrldq%20%20%20%20xmm1%2C%208%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2326.31%0D%0A%20%20%20%20%20%20%20%20psrldq%20%20%20%20xmm2%2C%2012%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2330.31%0D%0A%20%20%20%20%20%20%20%20pmovzxbd%20%20xmm0%2C%20xmm0%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2323.24%0D%0A%20%20%20%20%20%20%20%20pmovzxbd%20%20xmm1%2C%20xmm1%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2327.24%0D%0A%20%20%20%20%20%20%20%20pmovzxbd%20%20xmm2%2C%20xmm2%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2331.24%0D%0A%20%20%20%20%20%20%20%20movdqu%20%20%20%20XMMWORD%20PTR%20%5Brdx%5D%2C%20xmm3%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2333.36%0D%0A%20%20%20%20%20%20%20%20movdqu%20%20%20%20XMMWORD%20PTR%20%5B16%2Brdx%5D%2C%20xmm0%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2334.37%0D%0A%20%20%20%20%20%20%20%20movdqu%20%20%20%20XMMWORD%20PTR%20%5B32%2Brdx%5D%2C%20xmm1%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2335.37%0D%0A%20%20%20%20%20%20%20%20movdqu%20%20%20%20XMMWORD%20PTR%20%5B48%2Brdx%5D%2C%20xmm2%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2336.37%0D%0A%20%20%20%20%20%20%20%20add%20%20%20%20%20%20%20rdx%2C%2064%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2338.9%0D%0A%20%20%20%20%20%20%20%20cmp%20%20%20%20%20%20%20rax%2C%20rsi%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%2314.24%0D%0A%20%20%20%20%20%20%20%20jbe%20%20%20%20%20%20%20..B1.3%20%20%20%20&syntax=asIntel&uArchs=ICL&tools=uiCA&tools=OSACA&alignment=0)

So you should be a 4/16. cycles per character.