@lemire 2023-08-10
My numbers do not match yours...

```
git clone git@github.com:Nick-Nuon/simdutf.git
cd simdutf
git checkout sse_latin1_to_utf32
git clone git@github.com:lemire/unicode_lipsum.git
cmake -B build && cmake --build build
./build/benchmarks/benchmark -P convert_latin1_to_utf32 -F unicode_lipsum/wikipedia_mars/french.latin1.txt
```

I get...
```
convert_latin1_to_utf32+haswell, input size: 432305, iterations: 3000, dataset: unicode_lipsum/wikipedia_mars/french.latin1.txt
   0.563 ins/byte,    0.552 cycle/byte,    5.791 GB/s (1.5 %),     3.197 GHz,    1.020 ins/cycle
   0.563 ins/char,    0.552 cycle/char,    5.791 Gc/s (1.5 %)     1.00 byte/char
convert_latin1_to_utf32+icelake, input size: 432305, iterations: 3000, dataset: unicode_lipsum/wikipedia_mars/french.latin1.txt
   0.282 ins/byte,    0.521 cycle/byte,    5.942 GB/s (1.5 %),     3.096 GHz,    0.541 ins/cycle
   0.282 ins/char,    0.521 cycle/char,    5.942 Gc/s (1.5 %)     1.00 byte/char
convert_latin1_to_utf32+iconv, input size: 432305, iterations: 3000, dataset: unicode_lipsum/wikipedia_mars/french.latin1.txt
  32.024 ins/byte,    6.027 cycle/byte,    0.530 GB/s (0.4 %),     3.193 GHz,    5.314 ins/cycle
  32.024 ins/char,    6.027 cycle/char,    0.530 Gc/s (0.4 %)     1.00 byte/char
convert_latin1_to_utf32+icu, input size: 432305, iterations: 3000, dataset: unicode_lipsum/wikipedia_mars/french.latin1.txt
  54.798 ins/byte,   10.865 cycle/byte,    0.294 GB/s (4.8 %),     3.189 GHz,    5.044 ins/cycle
  54.798 ins/char,   10.865 cycle/char,    0.294 Gc/s (4.8 %)     1.00 byte/char
convert_latin1_to_utf32+westmere, input size: 432305, iterations: 3000, dataset: unicode_lipsum/wikipedia_mars/french.latin1.txt
   1.126 ins/byte,    0.495 cycle/byte,    6.461 GB/s (0.8 %),     3.197 GHz,    2.275 ins/cycle
   1.126 ins/char,    0.495 cycle/char,    6.461 Gc/s (0.8 %)     1.00 byte/char
``